requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
ccxt==1.50.88
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
cramjam==2.3.2
cryptography==3.4.7
dask==2021.5.1
distributed==2021.5.1
fastparquet==0.6.3
fsspec==2021.5.0
HeapDict==1.0.1
idna==2.10
locket==0.2.1
msgpack==1.0.2
multidict==5.1.0
multitasking==0.0.9
numpy==1.20.3
pandas==1.2.4
partd==1.2.0
psutil==5.8.0
pycares==4.0.0
pycparser==2.20
PyStore==0.1.22
python-dateutil==2.8.1
python-dotenv==0.17.1
python-snappy==0.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
thrift==0.13.0
toolz==0.11.1
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
zict==2.0.0
